Initialize system_gossip_message_validator module by extracting SystemGossipMessageValidator from system_executor
